Update release candidate
