+++ zkdb/.npmignore
@@ -3,6 +3,7 @@ src
 
 # Build test files
 build/src/api
+build/src/examples
 build/src/**/*.test.js
 build/src/**/*.test.map
 build/src/**/*.test.d.ts
@@ -32,3 +33,4 @@ jest.config.js
 # Never reveal your keys!
 keys
 .env
+*.log
