+++ zkdb/src/core/schema.ts
@@ -2,11 +2,53 @@ import { Struct, Poseidon, Field, InferProvable } from 'snarkyjs';
 import { BSON } from 'bson';
 export { Field } from 'snarkyjs';
 
+/**
+ * Interface for a schema
+ */
 export interface ISchema {
   serialize(): Uint8Array;
   hash(): Field;
 }
 
+/**
+ * Document schema is a wrapper for SnarkyJS Struct
+ * It provides the following additional methods:
+ * - serialize(): Serializes the document to a Uint8Array
+ * - hash(): Returns the hash of the document
+ * - decode(): Deserializes the document from a Uint8Array
+ *
+ * That makes it possible to use the document schema in the following way:
+ *
+ * ```ts
+ * // Define the schema of the document
+ * class Account extends Schema({
+ *   accountName: CircuitString,
+ *   balance: UInt32,
+ * }) {
+ *   // Deserialize the document from a Uint8Array
+ *   static deserialize(data: Uint8Array): Account {
+ *     return new Account(Account.decode(data));
+ *   }
+ *
+ *   // Index the document by accountName
+ *   index(): { accountName: string } {
+ *     return {
+ *       accountName: this.accountName.toString(),
+ *     };
+ *   }
+ *
+ *   // Serialize the document to a json object
+ *   json(): { accountName: string; balance: string } {
+ *     return {
+ *       accountName: this.accountName.toString(),
+ *       balance: this.balance.toString(),
+ *     };
+ *   }
+ * }
+ * ```
+ * @template T The inner type of struct of document
+ * @param _type The inner struct of document
+ */
 export type SchemaExtendable = <T>(_type: T) => Struct<
   T & InferProvable<T> & ISchema
 > & {
@@ -15,6 +57,7 @@ export type SchemaExtendable = <T>(_type: T) => Struct<
 
 export const Schema: SchemaExtendable = <A>(type: A) => {
   class Document extends Struct(type) {
+    // Serialize the document to a Uint8Array
     serialize(): Uint8Array {
       const anyThis = <any>this;
       const keys = Object.keys(type as any);
@@ -29,15 +72,16 @@ export const Schema: SchemaExtendable = <A>(type: A) => {
           throw new Error(`Cannot serialize ${key}`);
         }
       }
-
       return BSON.serialize(result);
     }
 
+    // Returns the hash of the document
     hash(): Field {
       return Poseidon.hash(Document.toFields(<any>this));
     }
   }
 
+  // Deserialize the document from a Uint8Array
   (Document as any).decode = (doc: Uint8Array): any => {
     const entires = Object.entries(<any>type);
     const docObj = BSON.deserialize(doc);
